Rename progress circle to progress spinner to make example work
