--- src/app/examples/progress-circle-overview/progress-circle-overview-example.ts
@@ -1,8 +1,0 @@
-import {Component} from '@angular/core';
-
-
-@Component({
-  selector: 'progress-circle-overview-example',
-  templateUrl: './progress-circle-overview-example.html',
-})
-export class ProgressSpinnerOverviewExample {}
