+++ src/app/examples/progress-spinner-configurable/progress-spinner-configurable-example.css
