+++ src/app/examples/progress-spinner-overview/progress-spinner-overview-example.css
