+++ src/app/examples/progress-spinner-overview/progress-spinner-overview-example.ts
@@ -1,0 +1,8 @@
+import {Component} from '@angular/core';
+
+
+@Component({
+  selector: 'progress-spinner-overview-example',
+  templateUrl: './progress-spinner-overview-example.html',
+})
+export class ProgressSpinnerOverviewExample {}
